+++ docs/src/components/DocPage/DocPage.tsx
@@ -9,7 +9,7 @@ interface DocPageProps {
 }
 
 const DocPage = ({ title, description, children }: DocPageProps) => (
-  <DocumentTitle title={`Stardust - ${title}`}>
+  <DocumentTitle title={`Fluent UI - ${title}`}>
     <div style={{ padding: '2rem', fontSize: '1.15rem', maxWidth: '100ch' }}>
       <Header as="h1" aria-level={2} content={title} description={description} align="center" />
       {children}