+++ packages/react/src/themes/teams/components/Icon/svg/icons/index.ts
@@ -145,7 +145,7 @@ import send from './send'
 import settings from './settings'
 import speakerMute from './speakerMute'
 import star from './star'
-import stardustCircle from './stardustCircle'
+import iconCircle from './iconCircle'
 import sticker from './sticker'
 import strike from './strike'
 import sync from './sync'
@@ -153,6 +153,7 @@ import tabs from './tabs'
 import table from './table'
 import tableAdd from './tableAdd'
 import tableDelete from './tableDelete'
+import tag from './tag'
 import teamCreate from './teamCreate'
 import teams from './teams'
 import toDoList from './toDoList'
@@ -178,7 +179,7 @@ import windowMinimize from './windowMinimize'
 import windowRestore from './windowRestore'
 
 export default {
-  'stardust-circle': stardustCircle,
+  'icon-circle': iconCircle,
   accept,
   add,
   'arrow-up': arrowUp,
@@ -215,8 +216,8 @@ export default {
   'canvas-add-page': canvasAddPage,
   chat,
   'chevron-down': chevronDown,
-  'stardust-chevron-start': chevronStart,
-  'stardust-chevron-end': chevronEnd,
+  'icon-chevron-start': chevronStart,
+  'icon-chevron-end': chevronEnd,
   'chevron-down-medium': chevronDownMedium,
   'chevron-right-medium': chevronRightMedium,
   'clipboard-copied-to': clipboardCopiedTo,
@@ -318,6 +319,7 @@ export default {
   table,
   'table-add': tableAdd,
   'table-delete': tableDelete,
+  tag,
   teams,
   thumbtack,
   'thumbtack-slash': thumbtackSlash,