+++ packages/react/src/utils/createComponent.tsx
@@ -1,22 +1,22 @@
 import { ReactAccessibilityBehavior, AccessibilityActionHandlers } from '@fluentui/react-bindings'
-import createComponentInternal, { CreateComponentReturnType } from './createComponent'
 import * as React from 'react'
 import * as _ from 'lodash'
 
+import createComponentInternal, { CreateComponentReturnType } from './createComponentInternal'
 import { ComponentSlotClasses, ComponentSlotStylesPrepared } from '../themes/types'
 import { ObjectOf } from '../types'
 
-export interface RenderStardustResultConfig {
+export interface CreateComponentRenderConfig {
   accessibility: ReactAccessibilityBehavior
   classes: ComponentSlotClasses
   rtl: boolean
   styles: ComponentSlotStylesPrepared
 }
 
-export interface CreateStardustComponentConfig<P> {
+export interface CreateComponentConfig<P> {
   displayName: string
   className?: string
-  render: (props: P & { stardust: RenderStardustResultConfig }) => React.ReactNode
+  render: (props: P & { config: CreateComponentRenderConfig }) => React.ReactNode
   defaultProps?: any
   actionHandlers?: AccessibilityActionHandlers
 }
@@ -27,13 +27,13 @@ const createComponent = <P extends ObjectOf<any> = any>({
   render,
   defaultProps,
   actionHandlers,
-}: CreateStardustComponentConfig<P>): CreateComponentReturnType<P> => {
+}: CreateComponentConfig<P>): CreateComponentReturnType<P> => {
   return createComponentInternal<P>({
     displayName,
     className,
     render: (config, props) => {
       const filteredConfig = _.pick(config, ['accessibility', 'classes', 'rtl', 'styles'])
-      return render(Object.assign({ stardust: filteredConfig }, props))
+      return render(Object.assign({ config: filteredConfig }, props))
     },
     defaultProps,
     actionHandlers,